separate out standalone from integration tests
