+++ stratigraph/__init__.py
@@ -1,4 +1,3 @@
-
 from stratigraph._version import get_versions
 __version__ = get_versions()['version']
 del get_versions
